requirements.txt
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
PyYAML==6.0
scikit_learn==1.2.2
tifffile==2023.2.3
torchsummary==1.5.1
torchvision==0.15.1
tqdm==4.64.1
zarr==2.14.1
